# [Refactor] Optimizations